♻ WV-25 refactor : api 함수 분리
